+++ gr/network-visualization/Datavisualization.Network/Layouts/ForceDirected/Layout/AbstractRenderer.vb
@@ -106,7 +106,7 @@ Namespace Layouts
         ''' </summary>
         ''' <param name="iTimeStep"><see cref="IForceDirected.Calculate(Single)"/></param>
         ''' <param name="physicsUpdate"></param>
-        Public Sub Draw(iTimeStep As Single, Optional physicsUpdate As Boolean = True) Implements IRenderer.Draw
+        Public Sub Draw(iTimeStep As Double, Optional physicsUpdate As Boolean = True) Implements IRenderer.Draw
             If physicsUpdate Then
                 ' 计算力的变化
                 Call forceDirected.Calculate(iTimeStep)
