+++ gr/network-visualization/Datavisualization.Network/Layouts/ForceDirected/Layout/Layout/ForceDirected2D.vb
@@ -50,7 +50,7 @@ Namespace Layouts
     ''' </summary>
     Public Class ForceDirected2D : Inherits ForceDirected(Of FDGVector2)
 
-        Public Sub New(iGraph As NetworkGraph, iStiffness As Single, iRepulsion As Single, iDamping As Single)
+        Public Sub New(iGraph As NetworkGraph, iStiffness As Double, iRepulsion As Double, iDamping As Double)
             MyBase.New(iGraph, iStiffness, iRepulsion, iDamping)
         End Sub
